// sgifd-frontend/src/app/models/ProjetProgrammeFinalise.ts
import {Exercice} from "./Exercice";
import {AxePrioritaire} from "./AxePrioritaire";
import {CategorieProjet} from "./CategorieProjet";
import {Cible} from "./Cible";
import {ConditionSuspensiveAccord} from "./ConditionSuspensiveAccord";
import {Envergure} from "./Envergure";
import {GrandSecteur} from "./GrandSecteur";
import {Localisation} from "./Localisation";
import {NiveauMaturite} from "./NiveauMaturite";
import {RessourceExterieure} from "./RessourceExterieure";
import {Structure} from "./Structure";
import {TypeRessourceInterieure} from "./TypeRessourceInterieure";
import {TypeCooperation} from "./TypeCooperation";
import {SousSecteur} from "./SousSecteur";
import {Secteur} from "./Secteur";
import {RessourceInterieureAnnuelle} from "./RessourceInterieureAnnuelle";
import {ConditionSuspensiveDecaissement} from "./ConditionSuspensiveDecaissement";
import {PieceIde} from "./Piece/PieceIde";
import {Piece} from "./Piece/Piece";
import {ProrogationProjet} from "./ProrogationProjet";
import {SecteurImpacte} from "./SecteurImpacte";
import {DocumentProgrammatique} from "./document-programmatique";

export class ProjetProgrammeFinalise {
  annee: Exercice;
  axePrioritaires: Array<AxePrioritaire>;
  categorieProjet: CategorieProjet;
  documentProgrammatique: DocumentProgrammatique;
  cibles: Array<Cible>;
  conditionSuspensiveAccords: Array<ConditionSuspensiveAccord>;
  contrePartieNationale: number;
  coutGlobalProjet: number;
  coutTotalRessourcesExterieures: number;
  createBy: string;
  createdAt: string;
  dateAchevementPrevue: string;
  dateApprobation: string;
  dateDemarrage: string;
  dateMiseEnVigueur: string;
  dateRactification: string;
  dateSignatureAccord: string;
  deleteBy: string;
  difficultes: string;
  dureeAnnees: Array<Exercice>;
  dureeProjet: number;
  envergure: Envergure;
  grandSecteur: GrandSecteur;
  id: number;
  libelle: string;
  localisations: Array<Localisation>;
  nbreProrogation: number;
  niveaumaturite: NiveauMaturite;
  objectifgeneral: string;
  objectifs: string;
  reference: string;
  ressourceExterieures: Array<RessourceExterieure>;
  ressourceInterieureAnnuelles: Array<RessourceInterieureAnnuelle>;
  secteur: Secteur;
  sousSecteur: SousSecteur;
  structureAgenceExecution: Array<Structure>;
  structureSousTutelle: Structure;
  type: string;
  typeCooperations: Array<TypeCooperation>;
  typeRessourceInterieures: Array<TypeRessourceInterieure>;
  close: boolean;
  valide: boolean;
  conditionSuspensiveDecaissements:Array<ConditionSuspensiveDecaissement>;
  files: Array<Piece>;
  prorogationProjets: Array<ProrogationProjet>;
  solutionEnvisagee: string;
  realisationAudit: string;
  secteurImpactes: Array<SecteurImpacte>;
  referencesMarche: string;
}